application assessment fields datatype corrected

+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application.Contracts/GrantApplications/GrantApplicationDto.cs
@@ -20,14 +20,14 @@ public class GrantApplicationDto : AuditedEntityDto<Guid>
     public string Category { get; set; } = string.Empty;
     public string EconomicRegion { get; set; } = string.Empty;
     public string City { get; set; } = string.Empty;
-    public double TotalProjectBudget { get; set; }
+    public decimal TotalProjectBudget { get; set; }
     public string Sector { get; set; } = string.Empty;
 
     public int AssessmentCount { get; set; } = 0;
     public int AssessmentReviewCount { get; set; } = 0;
 
     public string ProjectSummary { get; set; } = string.Empty;
-    public decimal TotalScore { get; set; } = 0;
+    public int TotalScore { get; set; } = 0;
     public decimal RecommendedAmount { get; set; } = 0;
     public decimal ApprovedAmount { get; set; } = 0;
     public string LikelihoodOfFunding { get; set; } = string.Empty;
